+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+### 21.10.9 [#1012](https://github.com/openfisca/openfisca-france/pull/1012)
+
+* Changement mineur.
+* Détails :
+  - Ajoute un test de performance sur les tests de CircleCI
+  - Ce test peut être lancé avec la commande: 
+    `python openfisca_france/scripts/performance_tests/test_circleci_builds.py 1717 1716`
+
 ### 21.10.8 [#1010](https://github.com/openfisca/openfisca-france/pull/1010)
 
 * Changement mineur.